+++ src/vws/exceptions/custom_exceptions.py
@@ -24,10 +24,3 @@ class TargetProcessingTimeout(Exception):
     """
     Exception raised when waiting for a target to be processed times out.
     """
-
-
-class ActiveMatchingTargetsDeleteProcessing(Exception):
-    """
-    Exception raised when a query is made with an image which matches a target
-    which has recently been deleted.
-    """